+++ threedigrid/admin/breaches/timeseries_mixin.py
@@ -1,8 +1,13 @@
+# (c) Nelen & Schuurmans.  GPL licensed, see LICENSE.rst.
+# -*- coding: utf-8 -*-
+from __future__ import unicode_literals
+from __future__ import print_function
+
+from threedigrid.admin.utils import combine_vars
 from threedigrid.orm.base.timeseries_mixin import ResultMixin
 from threedigrid.orm.base.fields import TimeSeriesArrayField
-from threedigrid.admin.constants import AGGREGATION_OPTIONS
-from threedigrid.admin.constants import BREACH_VARIABLES
-from threedigrid.admin.utils import combine_vars
+from threedigrid.orm.constants import AGGREGATION_OPTIONS
+from threedigrid.orm.constants import BREACH_VARIABLES
 
 
 class BreachResultsMixin(ResultMixin):
@@ -21,5 +26,5 @@ class BreachResultsMixin(ResultMixin):
         possible_vars = combine_vars(BREACH_VARIABLES, AGGREGATION_OPTIONS)
         possible_vars += BREACH_VARIABLES
         variables = set(possible_vars).intersection(netcdf_keys)
-        for var in variables:
-            setattr(self, var, TimeSeriesArrayField())
+        fields = {v: TimeSeriesArrayField() for v in variables}
+        self._meta.add_fields(fields, hide_private=True)