+++ threedigrid/admin/lines/timeseries_mixin.py
@@ -1,13 +1,23 @@
 # (c) Nelen & Schuurmans.  GPL licensed, see LICENSE.rst.
 # -*- coding: utf-8 -*-
 from threedigrid.orm.base.timeseries_mixin import ResultMixin
-from threedigrid.orm.base.fields import TimeSeriesArrayField
-from threedigrid.admin.constants import AGGREGATION_OPTIONS
-from threedigrid.admin.constants import LINE_VARIABLES
-from threedigrid.admin.utils import combine_vars
 
 
 class LineResultsMixin(ResultMixin):
+
+    class Meta:
+
+        # attributes for the given fields
+        field_attrs = ['units', 'long_name', 'standard_name']
+
+        composite_fields = {
+            'au': ['Mesh2D_au', 'Mesh1D_au'],
+            'u1': ['Mesh2D_u1', 'Mesh1D_u1'],
+            'q': ['Mesh2D_q', 'Mesh1D_q'],
+            '_mesh_id': ['Mesh2DLine_id', 'Mesh1DLine_id'],  # private
+        }
+
+        lookup_fields = ('id', '_mesh_id')
 
     def __init__(self, netcdf_keys, **kwargs):
         """Instantiate a line with netcdf results.
@@ -19,9 +29,3 @@ class LineResultsMixin(ResultMixin):
         :param kwargs:
         """
         super(LineResultsMixin, self).__init__(**kwargs)
-
-        possible_vars = combine_vars(LINE_VARIABLES, AGGREGATION_OPTIONS)
-        possible_vars += LINE_VARIABLES
-        variables = set(possible_vars).intersection(netcdf_keys)
-        for var in variables:
-            setattr(self, var, TimeSeriesArrayField())